Start trying to fix missing configuration argument
I'm totally ignoring the configuration map that's supplied to the init
part of the system creation. Instead I'm doing a complicated thing where
the caller can supply a list of args to be apply'd to the constructor.

This is a first attempt at mixing both, but I think the option to supply
the argument list should just go away. If you want to add config
options, put them in the tree.

This should play more nicely with that 'specify configuration' library
on which I was planning to base this.

Note that this version is pretty badly broken.