wip Start formalizing a context manager monad idea
One big question that remains open is "where does my system live?"

This is an attempt to answer that.

Really need to credit the blog post that credited the idea. I think the
author goes by the monikor softwareninja. Or maybe that was the core of
the domain name.

This version does not work, by any stretch of the imagination, but I
think it looks like a possibly useful solution to that issue.