Isolate context creation from the global
I'm pretty sure the context creation handler is a good idea. I have
serious misgivings about sticking it into a top-level var.

It should avoid most of the evils of global variables, but it still
makes me nervous to contemplate.

Still, in most cases there really will be only 1 that's active in a
given jvm. So maybe this isn't a horrible approach.