Cope with constructors w/ unknown specs
This might indicate a bug instead of something I should actually
handle. The problem turned up when I was trying to use a constructor
that didn't actually exist.

So this might be sweeping the actual problem under the rug. But the fix
was quick and easy, and it will be simple to revert.

Really just checking this in for the sake of getting repos merged.