Take baby-steps to debug nested creation failure
frereth-server fails to start because the event loop component it
inherits from frereth-common is getting set up to do extra
recursion. Trying to trace down why.

Starting by refactoring functions from pre-process to make it easier to
understand. I've already messed this up once and needed to revert, so
this is really just a milestone checkin in that direction.